requirements.txt
passlib
starlette
pydantic
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
insightface==0.7.3
scikit-learn==1.3.2
python-dotenv==1.0.0